Add .bash_aliases file
Alias `ls` to `eza` and `cat` to `bat`

Add function `dots` which will drop the user into fzf viewing the files
within the dotfiles repositories 'desktop' folder.
If a file is selected, it is opened in neovim if no file is chosen, or
for some reason the command errors, it will exit back to terminal.